requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==36.0.2
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-environ==0.8.1
django-mysql==4.5.0
django-on-heroku==1.1.2
django-rest-swagger==2.2.0
django-webtest==1.9.10
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.1
mysqlclient==2.1.0
openapi-codec==1.3.2
packaging==21.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
waitress==2.1.1
WebOb==1.8.7
WebTest==3.0.0
whitenoise==6.0.0
